@@ -10,12 +10,28 @@ declare(strict_types=1);
 
 namespace Oveleon\ContaoComponentStyleManager;
 
-use Symfony\Component\HttpKernel\Bundle\Bundle;
+use Symfony\Component\Config\Definition\Configurator\DefinitionConfigurator;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Symfony\Component\HttpKernel\Bundle\AbstractBundle;
 
-class ContaoComponentStyleManager extends Bundle
+class ContaoComponentStyleManager extends AbstractBundle
 {
-    public function getPath(): string
+    public function configure(DefinitionConfigurator $definition): void
     {
-        return \dirname(__DIR__);
+        $definition->rootNode()
+            ->children()
+                ->booleanNode('strict')
+                    ->defaultFalse()
+                ->end()
+            ->end()
+        ;
+    }
+
+    public function loadExtension(array $config, ContainerConfigurator $container, ContainerBuilder $builder): void
+    {
+        $container->import('../config/services.yaml');
+
+        $builder->setParameter('contao_component_style_manager.strict', $config['strict']);
     }
 }
