@@ -15,32 +15,16 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
-use Contao\ManagerPlugin\Routing\RoutingPluginInterface;
 use Contao\NewsBundle\ContaoNewsBundle;
 use Oveleon\ContaoComponentStyleManager\ContaoComponentStyleManager;
-use Symfony\Component\Config\Loader\LoaderResolverInterface;
-use Symfony\Component\HttpKernel\KernelInterface;
-use Symfony\Component\Routing\RouteCollection;
 
-class Plugin implements BundlePluginInterface, RoutingPluginInterface
+class Plugin implements BundlePluginInterface
 {
     public function getBundles(ParserInterface $parser): array
     {
         return [
-            BundleConfig::create(ContaoComponentStyleManager::class)
-                ->setLoadAfter([ContaoCoreBundle::class, ContaoCalendarBundle::class, ContaoNewsBundle::class])
-                ->setReplace(['componentStyleManager']),
+            new BundleConfig(ContaoComponentStyleManager::class)
+                ->setLoadAfter([ContaoCoreBundle::class, ContaoCalendarBundle::class, ContaoNewsBundle::class]),
         ];
-    }
-
-    /**
-     * @throws \Exception
-     */
-    public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel): RouteCollection|null
-    {
-        return $resolver
-            ->resolve('@ContaoComponentStyleManager/src/Controller')
-            ->load('@ContaoComponentStyleManager/src/Controller')
-        ;
     }
 }
